Add chapter on bug and feature branches
As part of the approval process for adding support for bug branches in
infra and devstack-gate, we agreed that we needed to document the
criteria for creating them and how they were meant to be used.

The same chapter is used to document the other type of exceptional
branch: feature branches.

Change-Id: I7fbd196711960db3d8e3cba643b5c6377801b3c9