Added project setup guide
This adds a new section, intended to capture common project setup
patterns in greater detail than the governance Consistent Testing
Interface. This change was suggested during discussion of the CTI
patch for npm-based testing available in this patch:
https://review.openstack.org/#/c/232756/

Python project guidelines are not yet included, contributions for
such are encouraged in a separate patch.

Change-Id: I00186c9af441c1778a22379634cdb2f918324bd8