stable: add hint about active maintenance and g-r caps
As we start to review requests to apply the stable:follows-policy
tag to projects, an easy way to know if a project broke stable
branches is if the project version is capped in global-requirements.

This is generally only possible starting with stable/liberty since
that's when we started using upper-constraints and capping in kilo
was by design (defensive rather than reactive).

Take for example Ifa9b07f7c4e93735aec33adb454e920c3c2d7141 which
proposes to cap python-troveclient<2.1.0 in stable/liberty g-r
because a backward incompatible change made to python-troveclient
in 2.1.0 breaks stable/liberty unit tests for trove.

This is a bit odd to think about with respect to stable branch policy,
because it's not actually the stable/liberty branch of
python-troveclient that is the issue, it's trunk python-troveclient
changes breaking stable branches of other projects.

But in this case, the stable maintenance team could use this cap
as justification for not applying the stable:follows-policy tag
to the python-troveclient project since it makes backward incompatible
changes which break stable branches.

Change-Id: I1fbf48168152ac0535d6fb0578200bcccdcb1a6d
Related-Bug: #1538506