Address PTL leave of absence
Based on TC discussion, we're going to make it clear that when
a community member is considering taking on the PTL role, they should
understand the amount of time they will need to dedicate. If they don't
have sufficient time for these items at any point of time due to
unexpected events, they need to take responsibility in handing things
off to someone to continue progress.

Change-Id: I9ac8e97631f84d427b81c72e8fd651b79a434268