Add link to Project Creator's Guide
It is quite easy for a newcomer who wants to create a new project
to mistakenly arrive at the Project Team Guide first before discovering
the Project Creator's Guide.  This is in part because:

  - the Project Team Guide is listed on http://docs.openstack.org/ above
    the link to the Infrastructure Documentation
    (http://docs.openstack.org/infra/),

  - the Infrastructure Manual is linked from within
    http://docs.openstack.org/infra/, so it's one extra hop away, and
    consequently

  - it is not immediately obvious from http://docs.openstack.org/ that
    the Infrastructure Documentation contains any info on project
    creation.

However even if it is practical to address these without compromising
other aspects of the design of the navigation around these front pages,
it would still be possible for newcomers to make the same mistake, e.g.
by following links from elsewhere.

This leads to reviews such as https://review.openstack.org/#/c/324798/.

Therefore make it clear at the top of the Project Team Guide that if you
don't already have a project created, you should go to the Creator's
Guide first.

Change-Id: I33c5d7ca110e0c2110e5fa6589cb4b953edafc0a
Signed-off-by: Adam Spiers <aspiers@suse.com>