update stable branch phase descriptions
The descriptions of stable branch phases here and on the
releases.openstack.org site do not match. This patch introduces some of
the terms used on the releases site, combined with the existing phase
numbering scheme, as a first step of unifying the two sets of
descriptions. A subsequent patch will update the releases site to link
to this section of the guide and more clearly identify the support phase
of legacy releases.

Change-Id: I92b1a80fcd8cd38d7e450269fe56ecc8ab6ba98f
Signed-off-by: Doug Hellmann <doug@doughellmann.com>