Add tips for writing good release notes
"Include a release note" has become a checkbox in some projects,
leading to every single patch having a reno note that vaguely rephrases
the git commit message or bug title, without considering how it may be
relevant to or make sense to a user not already familiar with the day
to day churn of the project.

The current guidelines explaining "when" a release note is needed are
useful, but don't seem sufficient to help people understand "how" to
write a good note that is both slightly more fleshed out than rewording
the git commit message title and also relevant to users.

This patch aims to be a starting point for being more explicit. If
there is existing documentation about this in OpenStack or elsewhere,
they would be good to reference as well.

Change-Id: I45c277d332ce8e670ef237122039445ca90477fe