Move 'big tent' to history
The words 'Big tent' have proven to be confusing to people.
Their initial meaning was that project teams can join OpenStack
is they are "one of us" (as defined by sharing our mission and
our principles). However, misinformed people (or people who
maliciously choose to be misinformed) took it to mean that
there is an absence of rules, and anything and anyone can call
themselves an OpenStack project.

This change describes the project structure reform from a
historical perspective, both to clarify what it actually meant
and to move it to past history. It also removes the last mention
of the 'big tent' that remained in the guide.

Change-Id: I00a9d22a16714c99cbf92f5556499c051d5a2d84