turn on warning-is-error
Sphinx warns because the type of the copyright value is not a string,
and with the flag on that prevents the build from proceeding.  I tried
using rst_epilog to add the text about where to contribute patches to
the bottom of each page, but the result makes it seem like it is part
of the text of the page. I think we want to add this as a feature to
the theme, to show up just like we do with the report-a-bug link in
the footer, so this patch removes the info under the assumption that
it will be handled by the theme.

Change-Id: Ia3e15b20a9d8a9e3598592b39dde316611c9323c
Signed-off-by: Doug Hellmann <doug@doughellmann.com>