Evolve project setup into a tech guidance chapter
Expand the short 'Project setup' chapter into a more complete
set of technical guidelines for projects, including pointers to
the list of base services, API guidelines...

This should be expanded in the future with other useful technical
design pointers.

Support a redirect from the old page name to the new page name.

Change-Id: Ic8e859dbe7261e15f4f6b3bcd84f31c39a8efb3c