Remove proactive backport stable instructions
This section of the stable branch guide gave an example of a specific
teams process for doing stable backports, using scripts from a repo that
no long exists. This is all useful information, but other teams are
handling this in a different way, and with some dead links to missing
scripts this may just cause more confusion than help.

Change-Id: I6a102a1aa50a917c03f1d81ebd846bdac1770311
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>