Remove setup.py check from pep8 recommendations
Using "python setup.py check -r -s" method of checking the package has
been deprecated with the new recommendation to build the sdist and
wheel, then running "twine check" against the output.

Luckily, there is already a job that covers this that only runs when the
README, setup.py, or setup.cfg files change, making running this in the
pep8 job redundant. This is covered by the test-release-openstack-python3
that is defined in the publish-to-pypi-python3 template that nearly all
repos use.

More details can be found in this mailing list post:

http://lists.openstack.org/pipermail/openstack-dev/2018-October/136136.html

Change-Id: I7e6f090a63d3a8d4790426931797189251b38de7
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>