Document use of the openstack-discuss mailing list
As the sun sets on our old openstack, openstack-dev,
openstack-operators and openstack-sigs mailing lists, provide a
durable reference to the recommended subject topic tags for the new
openstack-discuss mailing list which has supplanted them. This
initial set of tags was brainstormed at
https://etherpad.openstack.org/p/common-openstack-ml-topics but
subsequent additions or revisions should be made to this document
via code-review.

Change-Id: Iea0dc19775ac75a8979345e58e394094d30facd0