Add Project Update info to Events Section
I routinely have PTLs ask if they need to submit project
updates or onboarding to the CFP. They do not. Both
types of sessions are handled separately. This patch
adds directions about the project updates.

Since project onboarding will be different for Shanghai I did
not add that info in this patch. For Shanghai, the project
onboarding sessions will be a part of the PTG, where normally
they would be part of the Forum.

Change-Id: I8ef7cbd5fc81fd1e33e081bd460554a247e89bf5