Give an example of dealing with backport conflicts
The stable maint guide says to keep conflicts in the commit
message in a backport which is good but I frequently have to
give people an example of what that should look like.

This moves the conflicts part to its own sub-section for easy
linking and also links in an example of how to document
conflicts in a backport.

Change-Id: I550024dce50d19ac3dbbb6619159fafd8df8c314