Clarify deprecation tags and record deprecation cycle
Currently project deprecation is done by keeping the
project in project.yaml but with release-management as
deprecated which is not enough to now is it really
deprecated in governance or just release. Also we do not
record the deprecation time so that we can remove the project
after deprecation is over.

Let's add a new deprecation field with deprecation cycle recorded.

Depends-On: https://review.opendev.org/#/c/736084/

Change-Id: I2b86bfd801ad9d1df8d0ab2d8f269fb922366895