Document $series-last tag for Temepst to test EM branches
For testing for Extended Maintenance branches with branchless
Tempest and Tempest plugins, we need to release the new tag
for Tempest and its plugins as last compatible version for that
Extended Maintenance branches.

This commits adds the documentation about that tag $series-last
which will be used for branchless testing tooling.

Change-Id: I213112b9f03b79311c5e54200ed1cca0c8b817c3