Clarify optional lower bounds testing
Lower bounds testing was an optional thing to test but
projects raised question over maintaining it
- http://lists.openstack.org/pipermail/openstack-discuss/2021-January/019659.html
- http://lists.openstack.org/pipermail/openstack-discuss/2021-January/019672.html

During the discussion on ML and in the TC meeting, it is decided to
clarify that it is optional to test, and projects can drop it
if they want.
- http://eavesdrop.openstack.org/meetings/tc/2021/tc.2021-03-18-15.01.log.html#l-146

This commit makes few sentence updates about optional lower bounds testing.

Change-Id: I4b0e7d032845586720bfeb64761340ebc770d0e6