Migrate deprecation tag to a document here
This moves the guidelines of how feature deprecation should be handled
from the definition of assert:follows-standard-deprecation tag to
a document here. It reformats the verbiage therein to describe the
procedures as a guideline instead of adherence to an asserted tag. It
also modifies the procedures to be in line with the newly-adopted
TC resolution to modify the release cadence with tick and tock
designations which impact the timing of deprecations and feature
removals.

Change-Id: I9e6b252e81f893e0007c5b917328b73765d072c7