Update repository handling document
Sometime it is confusing whether to merge the
repo retirement/deprecation in governance and then
start the code/infra cleanup or other way around.
Usually all code/infra cleanup patches check if
there is patch submitted in governance or not.

If there is objection by TC or by any reasonTC decided
not to approve the repo retirement/deprecation then code/infra
cleanup including the project rename revert will be very costly.

To make it the process clear and in right sequence, updating the
document to merge the governance patch first and then start the
other steps.

Change-Id: I5c911f9dde8e4244156ad79898299dc3f0a11460