Explicitly mention the each steps of repo retirement
In repo retirement process, we need to wait for
governance patch to get TC agreement as well as to check
the proper retirement which need repo content removal first
to check the proper retirement. But as we refer to the OpenDev
manual for 3 steps together, this become confusing to mention the
right order. As per OpenStack process, those 3 steps need to be in
different order.

We have note about those but still it may confuse people to know
the right order.

I am extracting the each step explictly and what patch they need
to depends-on.

Change-Id: Ie932b38c79eaa5947959c6f642c041acfdb87679