Update retired project stable branch handling
When any project/repo is retired, we used to mark its
delivreables in stable branches as 'retired' so that release
team can use that flag to indentifiy the retired project.

Retirement means no maintenance for master as well as for
stable branches so it make sense to directly EOL all the
supported stable branches during retirement process.

Change-Id: Id6229dab375322b1026ffc8bbbb85058166f33a7