Move release cadence adjustment resolution to p-t-g
The TC passed a resolution for release cadence adjustment
- https://governance.openstack.org/tc/resolutions/20220210-release-cadence-adjustment.html

and resolution has the details content about the new process.
Sometime we need to evolve the process and docuement content
for more details. For that, moving the resolution content into
project-team-guide.

NOTE: this copy the resolution content as it is with modification
in title and replacing 'this resolution' to 'this document'. Any
improvement needed in content can be done in further changes.

Change-Id: I3242ff6c7759a34ebc1346a03a449a6219abc5f7